Organize optimization CLI modules
