+++ wmo/cli/optimize/__init__.py
@@ -1,0 +1,1 @@
+"""Optimization command modules for the WMO CLI."""
