+++ wmo/cli/optimize/router.py
@@ -10,9 +10,9 @@ import typer
 from rich.console import Console
 
 from wmo.cli.consent import can_prompt, require_spend_consent
+from wmo.cli.optimize.router_candidate_setup import collect_router_candidate_setup
 from wmo.cli.options import ROOT_OPTION, usage_error
 from wmo.cli.progress import progress_display
-from wmo.cli.router_candidate_setup import collect_router_candidate_setup
 from wmo.common.models import ProviderModelSelection, load_model_catalog
 from wmo.common.observability.telemetry import capture_completion_once
 from wmo.common.project import ProjectStore
