+++ wmo/cli/optimize/router_candidate_setup.py
